+++ packages/xlsx/src/XLSXWorksheet.js
@@ -1,21 +1,28 @@
 /**
  * Created by xmax on 16.11.2017.
  */
-const st = require('./tools')
+const tools = require('./tools')
 
 /**
  * Class XLSXWorksheet
  */
 class XLSXWorksheet {
   /**
-   *  @cfg {Object} margins
-   *  @cfg {Number} margins.left
-   *  @cfg {Number} margins.top
-   *  @cfg {Number} margins.right
-   *  @cfg {Number} margins.bottom
-   *  @cfg {Number} margins.header
-   *  @cfg {Number} margins.footer
-   *  all values are expressed in millimeters
+   * @constructor
+   * @param {Object} config
+   * @param {String} [config.title=Worksheet]
+   * @param {String} [config.name=Sheet] max length 30 symbols
+   * @param {String} [config.orientation=portrait]
+   * @param {Number} [config.worksheetScale] Scale of worksheet when printing
+   * @param {Boolean} [config.fixFirstColumn] Fix for unfilling first column
+   * @param {Object} [config.margins] all values are expressed in millimeters
+   * @param {Number} [config.margins.left]
+   * @param {Number} [config.margins.top]
+   * @param {Number} [config.margins.right]
+   * @param {Number} [config.margins.bottom]
+   * @param {Number} [config.margins.header]
+   * @param {Number} [config.margins.footer]
+   * @param workBook
    */
   constructor (config, workBook) {
     config = config || {}
@@ -23,6 +30,7 @@ class XLSXWorksheet {
     this.dataRows = []
     this.nextRowNum = 1
     this.columnProperies = ''
+    this.defStyles = []
     this.merge = []
     /**
      * Worksheet orientation
@@ -42,9 +50,9 @@ class XLSXWorksheet {
       minRowNum: 0, minColNum: 0, maxRowNum: 0, maxColNum: 0
     }
 
-    Object.assign(config, {
-      title: 'Worksheet', name: 'Лист', setActive: false, table: null, id: 1
-    })
+    config = Object.assign({
+      title: 'Worksheet', name: 'Sheet', setActive: false, table: null, id: 1
+    }, config)
 
     // todo check config
     Object.assign(this, config)
@@ -72,8 +80,8 @@ class XLSXWorksheet {
     d.maxRowNum = d.maxRowNum === 0 ? 1 : d.maxRowNum
 
     result += '<?xml version="1.0" encoding="UTF-8" standalone="yes"?>\n<worksheet xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships" xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" mc:Ignorable="x14ac" xmlns:x14ac="http://schemas.microsoft.com/office/spreadsheetml/2009/9/ac">' +
-      '<dimension ref="' + st.numAlpha(d.minColNum - (this.fixFirstColumn ? 1 : 0)) + d.minRowNum.toString() +
-      ':' + st.numAlpha(d.maxColNum - (this.fixFirstColumn ? 1 : 0)) + d.maxRowNum.toString() + '"/><sheetViews><sheetView ' + (this.setActive ? 'tabSelected="1" ' : '') +
+      '<dimension ref="' + tools.numAlpha(d.minColNum - (this.fixFirstColumn ? 1 : 0)) + d.minRowNum.toString() +
+      ':' + tools.numAlpha(d.maxColNum - (this.fixFirstColumn ? 1 : 0)) + d.maxRowNum.toString() + '"/><sheetViews><sheetView ' + (this.setActive ? 'tabSelected="1" ' : '') +
       ' workbookViewId="0"/></sheetViews><sheetFormatPr defaultRowHeight="15" />'
 
     result += this.columnProperies + '<sheetData>'
@@ -152,9 +160,9 @@ class XLSXWorksheet {
      */
 
     context.contentTypes.unshift('<Override PartName="/xl/worksheets/sheet' + this.id + '.xml" ContentType="application/vnd.openxmlformats-officedocument.spreadsheetml.worksheet+xml"/>')
-    context.props.unshift(st.escapeXML(this.name) || 'Sheet' + this.id)
+    context.props.unshift(tools.escapeXML(this.name) || 'Sheet' + this.id)
     context.xlRels.unshift('<Relationship Id="rId' + this.id + '" Type="http://schemas.openxmlformats.org/officeDocument/2006/relationships/worksheet" Target="worksheets/sheet' + this.id + '.xml"/>')
-    context.worksheets.unshift('<sheet name="' + (st.escapeXML(this.name) || 'Sheet' + this.id) + '" sheetId="' + this.id + '" r:id="rId' + this.id + '"/>')
+    context.worksheets.unshift('<sheet name="' + (tools.escapeXML(this.name) || 'Sheet' + this.id) + '" sheetId="' + this.id + '" r:id="rId' + this.id + '"/>')
   }
 
   /**
@@ -188,15 +196,16 @@ class XLSXWorksheet {
    */
 
   /**
-   * Add next row to Worksheet. Current row number can only increase.
+   * Add next row to Worksheet.
+   * _**Current row number can only increase!**_
    * For change the current row number, use {@link XLSX.csWorksheet#setRowNum}
    *
-   * @param {Array} config array of cells
-   * @param {Object} config.value type of Object must be in (string, boolean, date)
-   * @param {string} config.formula
-   * @param {Number} config.style style index for get call  WorkBook.style.add(congig)
-   * @param {Number} config.column column num
-   * @param {Array} [template] (optional) options for apply to config
+   * @param [{Object|*}] config Array of cell config {Object} If item type is not object it convert to yourValue >> {value: yourValue}
+   * @param {Object} cell.value type of Object must be (number, string, boolean, date)
+   * @param {string} [cell.formula]
+   * @param {Number|XLSXStyle|Object} [cell.style]  Style index for get call  WorkBook.style.add(congig)
+   * @param {Number} [cell.column] Column number First column 0
+   * @param {Array} [template] (optional) Template for apply to config. Array of {Cell}. Use this parameters for apply styles.
    * @param {Object} [rowConfig] (optional)
    * @param {Object} [rowConfig.height] (optional)
    * @return {Number}
@@ -243,18 +252,45 @@ class XLSXWorksheet {
 
     this.dataRows.push('> ') // + '" x14ac:dyDescent="0.25">';
 
-    var cellCount = 0
+    // let cellCount = 0
+    const hasTemplate = typeof template !== 'undefined' && Array.isArray(template) &&
+      template.length === config.length
+    const hasCellTemplate = (typeof rowConfig.cellTemplate !== 'undefined' && Array.isArray(rowConfig.cellTemplate) &&
+      rowConfig.cellTemplate.length === config.length)
+    const existCell = {}
+    config.forEach((cd) => {
+      existCell[cd.column] = true
+    })
+    // write def
+    let defR = this.defStyles[this.nextRowNum]
+    let lastColCheckDef = 0
 
     config.forEach((cell, index) => {
-      cellCount++
+      // cellCount++
       type = undefined
       valAsTagIs = false
-      // if (!Ext.isDefined(cell.value)){
-      //    return true;
-      // }
-      if (typeof template !== 'undefined' && typeof template[index] !== 'undefined') {
+      let cellType = typeof cell
+      if (cellType !== 'object' || (cell instanceof Date)) {
+        cell = {value: cell}
+      }
+      if (hasTemplate && typeof template[index] === 'object') {
         cell = Object.assign(cell, template[index])
       }
+      if (hasCellTemplate && typeof rowConfig.cellTemplate[index] === 'object') {
+        cell = Object.assign(cell, rowConfig.cellTemplate[index])
+      }
+      if (defR) { // write def styles before
+        for (let i = lastColCheckDef; i < cell.column; i++) {
+          if (defR[i]) {
+            this.dataRows.push(
+              '<c r="' + tools.numAlpha(i) + this.nextRowNum.toString() + '"' +
+              (' s="' + tools.extractId(defR[i]) + '"') + ' />'
+            )
+          }
+        }
+        lastColCheckDef = cell.column
+      }
+
       XLSXWorksheet.checkColNum(cell.column)
       if (typeof cell.value === 'string') {
         // Ext.String.htmlEncode
@@ -262,7 +298,7 @@ class XLSXWorksheet {
           cell.value = this.workBook.addString(cell.value)
           type = 's'
         } else {
-          cell.value = st.escapeXML(cell.value)
+          cell.value = tools.escapeXML(cell.value)
           valAsTagIs = true
           type = 'inlineStr'
         }
@@ -270,7 +306,7 @@ class XLSXWorksheet {
         cell.value = (cell.value ? 1 : 0)
         type = 'b'
       } else if (cell.value && cell.value instanceof Date) {
-        cell.value = st.convertDate(cell.value)
+        cell.value = tools.convertDate(cell.value)
         if (!cell.style) {
           cell.style = this.workBook.style.getDefDateStyle()
         }
@@ -282,8 +318,8 @@ class XLSXWorksheet {
         this.diapason.maxColNum = cell.column
       }
       this.dataRows.push(
-        '<c r="' + st.numAlpha(cell.column - (this.fixFirstColumn ? 1 : 0)) + this.nextRowNum.toString() + '"' +
-        (cell.style ? ' s="' + cell.style + '"' : '') +
+        '<c r="' + tools.numAlpha(cell.column - (this.fixFirstColumn ? 1 : 0)) + this.nextRowNum.toString() + '"' +
+        (cell.style ? ' s="' + tools.extractId(cell.style) + '"' : '') +
         (type ? ' t="' + type + '"' : '') +
         '>' +
         (cell.formula ? '<f>' + cell.formula + '</f>' : '') +
@@ -292,26 +328,56 @@ class XLSXWorksheet {
         '</c>'
       )
 
+      // remove used
+      if (this.defStyles[this.nextRowNum]) {
+        this.defStyles[this.nextRowNum][cell.column] = null
+      }
+
       if (cell.cellStyle) {
-        if (cell.cellStyle.rowSpan > 1) {
+        if (cell.cellStyle.rowSpan > 1 || cell.cellStyle.colSpan > 1) {
           this.addMerge({
-            colFrom: cellCount,
-            rowFrom: this.rowsCount,
-            colTo: cellCount,
-            rowTo: this.rowsCount + cell.cellStyle.rowSpan - 1
+            colFrom: cell.column,
+            rowFrom: this.nextRowNum,
+            colTo: cell.column + (cell.cellStyle.colSpan || 1) - 1,
+            rowTo: this.nextRowNum + (cell.cellStyle.rowSpan || 1) - 1
           })
         }
-        // colSpan
-        if (cell.cellStyle.colSpan > 1) {
-          this.addMerge({
-            colFrom: cellCount,
-            rowFrom: this.rowsCount,
-            colTo: cellCount + cell.cellStyle.colSpan - 1,
-            rowTo: this.rowsCount
-          })
+        if (cell.cellStyle.colSpan > 1 && cell.style && cell.style.config.border && cell.style.config.border.id > 0) {
+          // this.workBook.style.getStyle({border})
+          let cEnd = cell.column + cell.cellStyle.colSpan - 1
+          for (let i = cell.column + 1; i <= cEnd; i++) {
+            if (!existCell[i]) {
+              this.dataRows.push(
+                '<c r="' + tools.numAlpha(i) + this.nextRowNum.toString() + '"' +
+                (' s="' + tools.extractId(cell.style) + '"') + ' />'
+              )
+            }
+          }
+        }
+        if (cell.cellStyle.rowSpan > 1 && cell.style && cell.style.config.border && cell.style.config.border.id > 0) {
+          // remember style for down rows
+          let rwEnd = this.nextRowNum + cell.cellStyle.rowSpan - 1
+          let cEnd = cell.cellStyle.colSpan > 1 ? cell.column + cell.cellStyle.colSpan - 1 : cell.column
+          for (let ri = this.nextRowNum + 1; ri <= rwEnd; ri++) {
+            let rwCtx = this.defStyles[ri] = this.defStyles[ri] || []
+            for (let ci = cell.column; ci <= cEnd; ci++) {
+              rwCtx[ci] = cell.style
+            }
+          }
         }
       }
     })
+    // write def style fin
+    if (defR) {
+      defR.forEach((item, index) => {
+        if (!item || (index <= lastColCheckDef)) return
+        this.dataRows.push(
+          '<c r="' + tools.numAlpha(index) + this.nextRowNum.toString() + '"' +
+          (' s="' + tools.extractId(item) + '"') + ' />'
+        )
+      })
+      this.defStyles[this.nextRowNum] = null
+    }
 
     if (this.diapason.minRowNum === 0) {
       this.diapason.minRowNum = this.nextRowNum
@@ -332,15 +398,13 @@ class XLSXWorksheet {
     let res = []
     // this.columnProperties
     res.push('<cols>')
+    config = config.sort(f => f.column)
     config.forEach((column) => {
-      column.column = XLSXWorksheet.checkColNum(column.column)
-      res.push('<col min="')
-      res.push(column.column.toString()) // TODO: Индекс колонки с шириной
-      res.push('" max="')
-      res.push(column.column.toString())
-      res.push('" width="')
-      res.push(column.width.toString())
-      res.push('" customWidth="1"/>')
+      if (typeof column.column !== 'number' || typeof column.width !== 'number') {
+        throw new Error('Invalid parameters for XLSXWorksheet.setColsProperties')
+      }
+      let col = XLSXWorksheet.checkColNum(column.column) + 1
+      res.push(`<col min="${col}" max="${col}" width="${column.width}" customWidth="1"/>`)
     })
     res.push('</cols>')
     this.columnProperies = res.join('')
@@ -374,9 +438,9 @@ class XLSXWorksheet {
    *  add Merge for cells
    *  @param {Object|[colFrom, colTo, rowFrom, rowTo]} config
    *  @param {Number} config.colFrom
-   *  @param {Number} config.rowFrom
+   *  @param {Number} [config.rowFrom=current]
    *  @param {Number} config.colTo
-   *  @param {Number} config.rowTo
+   *  @param {Number} [config.rowTo=current]
    */
   addMerge (config) {
     if (Array.isArray(config)) {
@@ -393,8 +457,8 @@ class XLSXWorksheet {
     config.colTo = XLSXWorksheet.checkColNum(config.colTo)
     config.rowTo = XLSXWorksheet.checkRowNum(config.rowTo || this.nextRowNum)
 
-    this.merge.push('<mergeCell ref="' + st.numAlpha(config.colFrom - (this.fixFirstColumn ? 1 : 0)) + config.rowFrom +
-      ':' + st.numAlpha(config.colTo - (this.fixFirstColumn ? 1 : 0)) + config.rowTo + '"/>')
+    this.merge.push('<mergeCell ref="' + tools.numAlpha(config.colFrom - (this.fixFirstColumn ? 1 : 0)) + config.rowFrom +
+      ':' + tools.numAlpha(config.colTo - (this.fixFirstColumn ? 1 : 0)) + config.rowTo + '"/>')
     /*
      <mergeCells count="3">
      <mergeCell ref="A1:A3"/>