+++ packages/adminui-pub/CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [4.2.45]
+## Changed
+- `ubm_navshortcut` not not load `cmdData` attribute during startup, because this is CLOB,
+and fetching all CLOBS from table is very slow (at last for Oracle)
+
+### Added
+- Firing of BasePanel `beforeRefresh` event
+- ubdetailgrid: set allowBlank=false for row editor fields in case allowNull===false in meta files
+- ubcombobox: function getFieldValue - Get field value by name from fieldList
+- set `hideTrigger: true` by default for auto generated componets for numeric attributes
+
 ## [4.2.44]
 ## Fixed
 - set currency sign to '' for all languages to pass form validation for Currency attributes